refactor: move clickOutsideList into a hook
